% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-iowa_county_population.R
\docType{data}
\name{iowa_county_population}
\alias{iowa_county_population}
\title{Iowa population 2019}
\format{
A data frame with 99 rows and 6 variables:

\describe{
\item{fips}{FIPS code}
\item{county}{county name}
\item{population}{population}
\item{cumulative_population}{population in this and smaller counties}
\item{quantile_population}{proportion of state population in this and smaller counties}
\item{population_group}{group of counties such that each group has about a quarter of the state's population}
}
}
\source{
\href{https://www.icip.iastate.edu/tables/population/counties-estimates}{Iowa Community Indicators Program}
}
\usage{
iowa_county_population
}
\description{
Estimate of Iowa population by county, 2019.
}
\keyword{datasets}
